fixed small warning
git-svn-id: http://vtp.googlecode.com/svn/trunk@230 2ba57974-ef2e-0410-970c-f3cb55f6a86a

+++ TerrainSDK/vtdata/Content.cpp
@@ -203,7 +203,7 @@ void ContentVisitor::startElement (const char * name, const XMLAttributes &atts)
 			attval = atts.getValue("distance");
 			if (attval && *attval)
 			{
-				pModel->m_distance = atof(attval);
+				pModel->m_distance = (float)atof(attval);
 			}
 			else
 				pModel->m_distance = 0.0f;
